adds a products file in the controllers folder and adds a get all products controller
