+++ api/routes/products.js
@@ -4,6 +4,8 @@ const mongoose = require("mongoose");
 const multer = require('multer');
 const checkAuth = require('../middleware/check-auth');
 
+const ProductController = require('../controllers/products');
+
 const storage = multer.diskStorage({
   destination: function(req, file, cb) {
     cb(null, './uploads/');
@@ -32,41 +34,7 @@ const upload = multer({
 
 const Product = require("../models/product");
 
-router.get("/", (req, res, next) => {
-  Product.find()
-    .select("name price _id productImage")
-    .exec()
-    .then(docs => {
-      const response = {
-        count: docs.length,
-        products: docs.map(doc => {
-          return {
-            name: doc.name,
-            price: doc.price,
-            productImage: doc.productImage,
-            _id: doc._id,
-            request: {
-              type: "GET",
-              url: "http://localhost:3000/products/" + doc._id
-            }
-          };
-        })
-      };
-      //   if (docs.length >= 0) {
-      res.status(200).json(response);
-      //   } else {
-      //       res.status(404).json({
-      //           message: 'No entries found'
-      //       });
-      //   }
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json({
-        error: err
-      });
-    });
-});
+router.get("/", ProductController.products_get_all);
 
 router.post("/", checkAuth, upload.single('productImage'), (req, res, next) => {
   const product = new Product({
